Rotator file added
Now, the player moves up when the player presses the spacebar or the left mouse button. The player's color will be picked up randomly. A polygon collider is added to each slice of the circular disc. Every time the player collides with the disc, the collision is catched by the console.